refactor: remove AppSettings singleton
