+++ src/components/periodic-table-test/PeriodicTableTest.tsx
@@ -1,8 +1,8 @@
 import classNames from "classnames";
 import * as React from "react";
 import { useHistory } from "react-router-dom";
-import { IPeriodicTableTestSettings } from "@/AppSettings";
 import { Element } from "@/Element";
+import { ElementsSettings } from "@/hooks/useAppSettings";
 import { useElements } from "@/hooks/useElements";
 import { useLocale } from "@/hooks/useLocale";
 import { TEST_SELECTION } from "@/routes";
@@ -29,7 +29,7 @@ function PeriodicTableTest() {
   const { settings, updateSettings } = usePeriodicTableTestSettings();
 
   const createTestQuestions = React.useCallback(
-    (settings: IPeriodicTableTestSettings) => {
+    (settings: ElementsSettings) => {
       if (!settings.elements) {
         return [];
       }
