+++ src/components/periodic-table-test/hooks/usePeriodicTableTestSettings.ts
@@ -1,6 +1,5 @@
 import * as React from "react";
-import { IPeriodicTableTestSettings } from "@/AppSettings";
-import { useAppSettings } from "@/hooks/useAppSettings";
+import { useAppSettings, ElementsSettings } from "@/hooks/useAppSettings";
 import { useElements } from "@/hooks/useElements";
 
 export function usePeriodicTableTestSettings() {
@@ -25,7 +24,7 @@ export function usePeriodicTableTestSettings() {
   const storedSettings = settings.tests.periodicTable;
 
   const updateTestSettings = React.useCallback(
-    (updateFunction: (settings: IPeriodicTableTestSettings) => void) => {
+    (updateFunction: (settings: ElementsSettings) => void) => {
       updateSettings((settings) => {
         updateFunction(settings.tests.periodicTable);
       });
