--- src/defaultSettings.ts
@@ -1,18 +1,0 @@
-import { Settings } from "./AppSettings";
-import { getBrowserLocale } from "./hooks/useLocale";
-import { DEFAULT_THEME } from "./hooks/useTheme";
-
-const defaultSettings: Settings = {
-  locale: getBrowserLocale(),
-  tests: {
-    periodicTable: {
-      elements: null,
-    },
-    valences: {
-      elements: null,
-    },
-  },
-  theme: DEFAULT_THEME,
-};
-
-export default defaultSettings;
