+++ src/setupTests.ts
@@ -1,5 +1,3 @@
 import "@testing-library/jest-dom";
 import "@testing-library/jest-dom/extend-expect";
 import "whatwg-fetch";
-import AppSettings from "./AppSettings";
-AppSettings.loadSettings();
